No missing value for class 31 descriptors
The missing value handling should not apply to class 31 descriptors.
Quote from the WMO BUFR Spec:

```
Missing values shall be set to fields of all ones (e.g. each octet shall
be set to 11111111 binary). This shall apply to code tables as well as
data elements; flag tables shall be augmented to contain a missing
indicator bit where this is deemed to be necessary. This regulation does
not apply to the data description operator qualifiers in Class 31 of
Table B.
```

Resolves: #44